Final Project (#4)
Summary: Implements the home view which shows the Desserts and could be used to display any other categories. The initial meals are sorted alphabetically. There is an option to also sort in descending order as well as search desserts. The detail view is also implemented and will show other details such as Instructions and the ingredients and measurements for the meal, as well as a link that will take you to YouTube if applicable. Unit tests were created to ensure outcome is expected. The project was done in both UIKit in one view and SwiftUI in the other, on purpose. Thanks for reviewing!